Version 1.1.0
* Add mean shift filter algorithm.
* Now the time-consuming algorithms will be quickly interrupted after
the main panel is closed, to avoid the situation that the program still
remains in the background after the window is closed.
* Now the mouse on the thumbnails of the processing results will display
the original image, right click on the image can be copied or saved.
* Removed "Export Result" button.
* Now when a algorithm is running, the corresponding progress bar will
be displayed on the lower right of the main panel, and the cursor will
be changed to "busy".
* Adjusted the criterion of "connectivity-based correction" to improve
the accuracy of error correction.
* Some other technical adjustments.